+++ database/migrations/2021_12_05_000003_create_subscriptions_table.php
@@ -3,16 +3,15 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
-use FintechSystems\Payfast\Enums\PaymentMethod;
 
-class CreateSubscriptionsTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('subscriptions', function (Blueprint $table) {
             $table->id();
@@ -25,31 +24,12 @@ class CreateSubscriptionsTable extends Migration
             $table->timestamp('trial_ends_at')->nullable();
             $table->timestamp('paused_from')->nullable();
             $table->timestamp('ends_at')->nullable();
-            $table->timestamp('next_bill_at')->nullable(); // Custom - added
-            $table->timestamp('cancelled_at')->nullable(); // Custom - added
-            $table->string('payment_method')->nullable(); // Custom - added
-            $table->timestamps();
-
-            $table->index(['billable_id', 'billable_type']);
-
-            // Fields added post design
+            $table->timestamp('next_bill_at')->nullable();
+            $table->timestamp('cancelled_at')->nullable();
+            $table->string('payment_method')->nullable();
             $table->string('merchant_payment_id')->nullable();
-
-            // $table->unsignedBigInteger('amount')->nullable();
-            // $table->unsignedInteger('cycles')->nullable();
-            // $table->unsignedInteger('cycles_complete')->nullable();
-            // $table->unsignedInteger('frequency')->nullable();
-
-            // $table->unsignedInteger('payfast_status')->nullable();
-            // $table->string('payfast_status_reason')->nullable();
-            // $table->string('payfast_status_text')->nullable();
-
-            // $table->string('status');
-
-            // $table->string('payment_status')->nullable();
-            // $table->string('subscription_status')->nullable();
-
-            // $table->timestamp('next_run_at')->nullable();
+            $table->timestamps();
+            $table->index(['billable_id', 'billable_type']);
         });
     }
 
@@ -58,8 +38,8 @@ class CreateSubscriptionsTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('subscriptions');
     }
-}
+};