+++ database/migrations/2021_12_05_000004_create_orders_table.php
@@ -4,14 +4,14 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateOrdersTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
@@ -19,7 +19,6 @@ class CreateOrdersTable extends Migration
             $table->string('billable_type');
             $table->string('ip_address');
             $table->timestamps();
-
             $table->index(['billable_id', 'billable_type']);
         });
     }
@@ -29,8 +28,8 @@ class CreateOrdersTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('orders');
     }
-}
+};